Quick patch to enable nrf9151 support in embassy-net-nrf91
Patched `embassy-net-nrf91` to work on the nRF9151.
A feature flag to select between the nRF9151 and nRF9160 has been
introduced.
The current patch for the nRF9151 is a hot-fix, and I make no guarantees
as to the quality and reliability of the fix, initial testing does
however show it working.

The example from nRF9160 has been copied to the nRF9151 example folder and
has been tested on an nRF9151-DK and is confirmed working.
The example nRF9160 example has been run on an nRF9160-DK and is confirmed
working after modifications.